panning behaviour in event-driven mode works
